+++ src/main/java/com/clay/wmp/user/service/UserService.java
@@ -1,15 +1,17 @@
 package com.clay.wmp.user.service;
 
 import com.clay.wmp.common.exception.DuplicateResourceException;
+import com.clay.wmp.common.exception.ResourceInUseException;
 import com.clay.wmp.common.exception.ResourceNotFoundException;
-import com.clay.wmp.user.dto.RegisterUserRequest;
-import com.clay.wmp.user.dto.UpdateRoleDto;
-import com.clay.wmp.user.dto.UserDto;
-import com.clay.wmp.user.dto.UserMapper;
+import com.clay.wmp.project.service.ProjectService;
+import com.clay.wmp.task.service.TaskService;
+import com.clay.wmp.team.service.TeamService;
+import com.clay.wmp.user.dto.*;
 import com.clay.wmp.user.entity.User;
 import com.clay.wmp.user.repository.UserRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -23,42 +25,57 @@ public class UserService {
 
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
+    private final ProjectService projectService;
+    private final TaskService taskService;
+    private final TeamService teamService;
 
-    public UserService(UserRepository userRepository, PasswordEncoder passwordEncoder) {
+    public UserService(UserRepository userRepository, PasswordEncoder passwordEncoder,
+                       ProjectService projectService, TaskService taskService, TeamService teamService) {
         this.userRepository = userRepository;
         this.passwordEncoder = passwordEncoder;
+        this.projectService = projectService;
+        this.taskService = taskService;
+        this.teamService = teamService;
     }
 
     @Transactional(readOnly = true)
     public List<UserDto> getAllUsers() {
         return userRepository.findAll()
-                .stream().map(UserMapper::mapToDto).toList();
+                .stream().map(UserDto::fromUser).toList();
     }
 
     @Transactional(readOnly = true)
     public UserDto getUserById(Long id) {
         return userRepository.findById(id)
-                .map(UserMapper::mapToDto)
+                .map(UserDto::fromUser)
                 .orElseThrow(() -> new ResourceNotFoundException("User not found"));
     }
 
     @Transactional(readOnly = true)
     public UserDto getUserByUsername(String username) {
         return userRepository.findByUsernameIgnoreCase(username)
-                .map(UserMapper::mapToDto)
+                .map(UserDto::fromUser)
                 .orElseThrow(() ->  new ResourceNotFoundException("User not found"));
     }
 
     @Transactional(readOnly = true)
     public UserDto getUserByEmail(String email) {
         return userRepository.findByEmailIgnoreCase(email)
-                .map(UserMapper::mapToDto)
+                .map(UserDto::fromUser)
                 .orElseThrow(() ->  new ResourceNotFoundException("User not found"));
     }
 
     @Transactional(readOnly = true)
     public List<UserDto> getUsersByRole(User.UserRole role) {
-        return userRepository.findByRole(role).stream().map(UserMapper::mapToDto).toList();
+        return userRepository.findByRole(role).stream().map(UserDto::fromUser).toList();
+    }
+
+    @Transactional(readOnly = true)
+    public UserDto getCurrentUser() {
+        Long id = 1L; // TODO get current user id from auth
+        return userRepository.findById(id)
+                .map(UserDto::fromUser)
+                .orElseThrow(() -> new ResourceNotFoundException("User not found"));
     }
 
     @Transactional(readOnly = true)
@@ -75,57 +92,77 @@ public class UserService {
     @Transactional
     public UserDto createUser(RegisterUserRequest registerUserRequest) {
 
-        if(userRepository.existsByUsername(registerUserRequest.username())) {
+        if(userRepository.existsByUsernameIgnoreCase(registerUserRequest.username())) {
             throw new DuplicateResourceException("Username already exists");
         }
 
-        if(userRepository.existsByEmail(registerUserRequest.email())) {
+        if(userRepository.existsByEmailIgnoreCase(registerUserRequest.email())) {
             throw new DuplicateResourceException("Email already exists");
         }
 
+        // TODO we will need to add security to the password rules later, as well as force change.
         var newUser = new User(
                 registerUserRequest.username(),
                 registerUserRequest.name(),
                 registerUserRequest.email(),
                 passwordEncoder.encode(registerUserRequest.password())
         );
-        return UserMapper.mapToDto(userRepository.save(newUser));
+        return UserDto.fromUser(userRepository.save(newUser));
     }
 
     @Transactional
-    public UserDto updateUser(Long id, UserDto userDto) {
+    public UserDto updateUser(Long id, UpdateUserProfileRequest updateUserProfileRequest) {
         var user = userRepository.findById(id)
-                .orElseThrow(RuntimeException::new);
+                .orElseThrow(() -> new ResourceNotFoundException("User not found"));
 
-        if(!user.getUsername().equalsIgnoreCase(userDto.username()) &&
-                userRepository.existsByUsername(userDto.username())) {
+        if(!user.getUsername().equalsIgnoreCase(updateUserProfileRequest.username()) &&
+                userRepository.existsByUsernameIgnoreCase(updateUserProfileRequest.username())) {
             log.debug("Username already exists, throwing exception");
             throw new DuplicateResourceException("Username already exists");
         }
 
-        if(!user.getEmail().equalsIgnoreCase(userDto.email()) &&
-                userRepository.existsByEmail(userDto.email())) {
+        if(!user.getEmail().equalsIgnoreCase(updateUserProfileRequest.email()) &&
+                userRepository.existsByEmailIgnoreCase(updateUserProfileRequest.email())) {
             throw new DuplicateResourceException("Email already exists");
         }
 
-        user.setUsername(userDto.username());
-        user.setName(userDto.name());
-        user.setEmail(userDto.email());
-        return UserMapper.mapToDto(userRepository.save(user));
+        user.setUsername(updateUserProfileRequest.username());
+        user.setName(updateUserProfileRequest.name());
+        user.setEmail(updateUserProfileRequest.email());
+        return UserDto.fromUser(userRepository.save(user));
     }
 
-    public UserDto updateRole(Long id, UpdateRoleDto updateRoleDto) {
+    @Transactional
+    public UserDto updateRole(Long id, UpdateUserRoleRequest updateUserRoleRequest) {
         var user = userRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("User not found"));
 
-        user.setRole(updateRoleDto.role());
-        return UserMapper.mapToDto(userRepository.save(user));
+        user.setRole(updateUserRoleRequest.role());
+        return UserDto.fromUser(userRepository.save(user));
+    }
+
+    @Transactional
+    public UserDto updatePassword(Long id, UpdateUserPasswordRequest updateUserPasswordRequest) {
+        var user = userRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("User not found"));
+
+        if (!user.getPassword().equals(passwordEncoder.encode(updateUserPasswordRequest.currentPassword()))) {
+            throw new BadCredentialsException("Incorrect credentials"); //might need a custom bad credentials exception
+        }
+        //New Password checks should be added before just allowing an update.
+        //validatePassword(updateUserPasswordRequest.password());
+        user.setPassword(passwordEncoder.encode(updateUserPasswordRequest.newPassword()));
+        return UserDto.fromUser(userRepository.save(user));
     }
 
     @Transactional
     public void deleteUser(Long id) {
-        if(!userRepository.existsById(id)) {
+        if (!userRepository.existsById(id)) {
             throw new ResourceNotFoundException("User not found");
+        }
+        // DB Constraints can handle this but going through services is future consideration for microservices
+        if (projectService.userHasProjects(id) || teamService.isTeamMember(id) || taskService.hasTasks(id)) {
+            throw new ResourceInUseException("User cannot be deleted due to associated records");
         }
         userRepository.deleteById(id);
     }
